Refactor location form
