--- AgentDiscoveries-Frontend/app/src/components/admin/location-form.jsx
@@ -1,158 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import {Button, ControlLabel, Form, FormControl, FormGroup} from 'react-bootstrap';
-import {GoogleApiWrapper, Map, Marker} from 'google-maps-react';
-import {StyledDegreeDirection, DegreeWrapper, FormRow, StyledFormGroup, MapWrapper} from "./location-form.style";
-import {apiGet, apiPost, apiPut} from '../utilities/request-helper';
-import Message from '../message';
-
-const DegreeDirection = ({ isNorth }) => (
-    <StyledDegreeDirection>
-        <DegreeWrapper>&deg;</DegreeWrapper>
-        <span>{isNorth ? 'N' : 'W'}</span>
-    </StyledDegreeDirection>
-);
-
-const LocationForm = ({ id, google }) => {
-    const [isLoaded, setIsLoaded] = useState(false);
-
-    const [values, setValues] = useState({
-        siteName: '',
-        location: '',
-        timeZone: '',
-        regionId: '',
-        latitude: '',
-        longitude: ''
-    });
-
-    const { siteName, location, timeZone, regionId, latitude, longitude } = values;
-
-    const [message, setMessage] = useState({});
-
-    useEffect(() => {
-        (async () => {
-            if (id) await loadLocation(id);
-            setIsLoaded(true);
-        })();
-    }, []);
-
-
-    const handleChange = event => {
-        const { name, value } = event.target;
-        setValues({ ...values, [name]: value });
-    }
-
-    const handleSubmit = async event => {
-        event.preventDefault();
-
-        const body = {
-            siteName,
-            location,
-            timeZone,
-            regionId: parseInt(regionId),
-            latitude: parseFloat(latitude),
-            longitude: parseFloat(longitude)
-        };
-
-        try {
-            if (id) {
-                await apiPut('locations', body, id);
-            } else {
-                await apiPost('locations', body);
-            }
-
-            window.location.hash = '#/admin/locations';
-        } catch (error) {
-            setMessage({ message: error.message, type: 'danger' })
-        }
-    }
-
-
-    const loadLocation = async () => {
-        try {
-            setValues({...values, ...await apiGet('locations', id)});
-        } catch (error) {
-            setMessage({ message: error.message, type: 'danger' });
-        }
-    };
-
-    if (!isLoaded) return <span>Loading...</span>;
-
-    return (
-        <div className='col-md-8 col-md-offset-2'>
-            <Message message={message} />
-            <div className='col-md-12'>
-                <Form onSubmit={handleSubmit}>
-                    <h3>{id ? 'Edit' : 'Create'} Location</h3>
-
-                    <FormGroup>
-                        <ControlLabel>Site Name</ControlLabel>
-                        <FormControl type='text' required
-                                     placeholder='Enter site name'
-                                     name='siteName'
-                                     value={siteName}
-                                     onChange={handleChange}/>
-                    </FormGroup>
-                    <FormGroup>
-                        <ControlLabel>Location Name</ControlLabel>
-                        <FormControl type='text' required
-                                     placeholder='Enter location name'
-                                     name='location'
-                                     value={location}
-                                     onChange={handleChange}/>
-                    </FormGroup>
-                    <FormGroup>
-                        <ControlLabel>Time Zone</ControlLabel>
-                        <FormControl type='text' required
-                                     placeholder='Enter time zone (e.g. "Europe/London")'
-                                     name='timeZone'
-                                     value={timeZone}
-                                     onChange={handleChange}/>
-                    </FormGroup>
-                    <FormGroup>
-                        <ControlLabel>Region</ControlLabel>
-                        <FormControl type='number'
-                                     placeholder='Enter region ID (optional)'
-                                     name='regionId'
-                                     value={regionId}
-                                     onChange={handleChange}/>
-                    </FormGroup>
-                    <FormRow>
-                        <StyledFormGroup>
-                            <ControlLabel>Latitude</ControlLabel>
-                            <FormRow alignCenter>
-                                <FormControl type='number'
-                                             placeholder='00.000'
-                                             name='latitude'
-                                             value={latitude}
-                                             onChange={handleChange}/>
-                                <DegreeDirection isNorth />
-                            </FormRow>
-                        </StyledFormGroup>
-                        <StyledFormGroup>
-                            <ControlLabel>Longitude</ControlLabel>
-                            <FormRow alignCenter>
-                                <FormControl type='number'
-                                             placeholder='00.000'
-                                             name='longitude'
-                                             value={longitude}
-                                             onChange={handleChange}/>
-                                <DegreeDirection />
-                            </FormRow>
-                        </StyledFormGroup>
-                    </FormRow>
-                    <MapWrapper>
-                        <Map google={google}
-                             initialCenter={{lat: parseFloat(latitude), lng: -parseFloat(longitude)}}
-                             center={{lat: parseFloat(latitude), lng: -parseFloat(longitude)}}
-                             zoom={14}>
-                            <Marker title={siteName} name='Location' position={{lat: parseFloat(latitude), lng: -parseFloat(longitude)}}/>
-                        </Map>
-                    </MapWrapper>
-                    <Button type='submit'>Submit</Button>
-                </Form>
-            </div>
-        </div>
-    );
-}
-
-export default GoogleApiWrapper({apiKey: 'GOOGLE_MAPS_API_KEY'})(LocationForm);
