+++ AgentDiscoveries-Frontend/app/src/components/admin/location/degree-direction.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+import {StyledDegreeDirection, DegreeWrapper} from './degree-direction.style';
+
+export default ({ isNorth }) => (
+    <StyledDegreeDirection>
+        <DegreeWrapper>&deg;</DegreeWrapper>
+        <span>{isNorth ? 'N' : 'W'}</span>
+    </StyledDegreeDirection>
+);
