+++ AgentDiscoveries-Frontend/app/src/components/admin/location/degree-direction.style.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const StyledDegreeDirection = styled.div`
+    display: flex;
+    font-size: 16px;
+`;
+
+export const DegreeWrapper = styled.span`
+    margin: 0 4px 0 2px;
+`;
