+++ AgentDiscoveries-Frontend/app/src/components/admin/location/google-map.style.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const MapWrapper = styled.div`
+    position: relative;
+    height: 200px;
+    width: 100%;
+    margin-bottom: 15px;
+`;
