+++ AgentDiscoveries-Frontend/app/src/components/admin/location/index.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import Form from './form';
+
+export default ({ id }) => (
+    <div className='col-md-8 col-md-offset-2'>
+        <Form id={id} />
+    </div>
+);
