+++ AgentDiscoveries-Frontend/app/src/components/app.jsx
@@ -14,7 +14,7 @@ import LocationReportSubmit from './submit-forms/submit-location-report';
 import RegionSummarySubmit from './submit-forms/submit-region-summary';
 import Entities from './admin/entities-view';
 import TodaysCodePage from './todays-code-page';
-import LocationForm from './admin/location-form';
+import Location from './admin/location';
 import RegionForm from './admin/region-form';
 import UserForm from './admin/user-form';
 import Error from './error';
@@ -41,11 +41,11 @@ const App = () => {
                     <Route path='/admin/regions' exact render={() => <Page><Entities api='regions' key='regions'/></Page>} />
                     <Route path='/admin/users' exact render={() => <Page><Entities api='users' key='users'/></Page>} />
 
-                    <Route path='/admin/locations/add' render={() => <Page><LocationForm/></Page>} />
+                    <Route path='/admin/locations/add' render={() => <Page><Location/></Page>} />
                     <Route path='/admin/regions/add' render={() => <Page><RegionForm/></Page>} />
                     <Route path='/admin/users/add' render={() => <Page><UserForm/></Page>} />
 
-                    <Route path='/admin/locations/edit/:id' render={props => <Page><LocationForm id={props.match.params.id} /></Page>} />
+                    <Route path='/admin/locations/edit/:id' render={props => <Page><Location id={props.match.params.id} /></Page>} />
                     <Route path='/admin/regions/edit/:id' render={props => <Page><RegionForm id={props.match.params.id} /></Page>} />
                     <Route path='/admin/users/edit/:id' render={props => <Page><UserForm id={props.match.params.id} /></Page>} />
