+++ AgentDiscoveries-Frontend/node_tmp/node-v12.16.1-darwin-x64/share/man/man1/node.1
@@ -111,7 +111,7 @@ Requires Node.js to be built with
 .Sy ./configure --openssl-fips .
 .
 .It Fl -enable-source-maps
-Enable experimental Source Map V3 support for stack traces.
+Enable experimental Source GoogleMap V3 support for stack traces.
 .
 .It Fl -experimental-json-modules
 Enable experimental JSON interop support for the ES Module loader.
